converted sketch and union to templates (allocator is not used yet), moved implementation out, return sketch by value
